requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dropbox==11.33.0
elastic-transport==8.1.2
elasticsearch==8.3.3
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
ply==3.11
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
stone==3.3.1
tika==1.24
typing-extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
